[FIX] point_of_sale: use client validation date instead of creation date
9f541d29bf58c3604f6c834ba25e1625c9ad1480 made it so that date_order on
pos.order records is set using the creation_date of the clientside Order
record to avoid weird date_order values when syncing to the server is
not done immediately.

The downside of this approach is that creation_date is set during Order
creation on the client. A user that has an open POS will also have an
open order. So if it takes one hour before the first order can be made
then the date_order will reflect this.

To avoid this we will instead use a validation date, which will be set
when the user clicks on the green 'Validate' button.

Because this is a stable release we won't change the name of the key in
the JSON data that gets sent to the server. This means that from the
server point of view nothing changes.

Fixes #11933