[FIX] product: pricelist based on supplier price
In the case of a purchase pricelist based on the supplier price, the
price selected is incorrect if the supplier is not in the list of
suppliers of the product. Indeed, in this case, the price of the first
supplier is chosen.

We also need to make sure to go through all the pricelist items found.

opw-678440