[FIX] web: set title on Many2Many tags
When the content of a Many2Many tag is too large, it is truncated. This
is an issue since there is no easy way to see it.

The fix adds a title on the tag with the content.

X-port of commit odoo/enterprise@bd3bfe04221b685d4e4eb7dbf000c4cd8b0b9d69