[FIX] report: correctly handle broken wkhtmltopdf
It's possible that the command `wkhtmltopdf --version` returns an empty
string (in the use case, it output later in stdout "wkhtmltopdf: error
while loading shared libraries: libfontconfig.so.1: cannot open shared
object file: No such file or directory").

We chose to display the report in html in this case and also to show
a warning to the user inviting him to fix his installation.